pyproject.toml
[tool.poetry]
name = "demo-app"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.8"
uvicorn = {extras = ["standard"], version = "^0.34.0"}
sqlalchemy = "^2.0.38"
pymysql = "^1.1.1"
pyjwt = "^2.10.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
